// src/app/models/admin/dangkyhoatdongkhoangsan/dktanthugiahan.model.ts
export class OutputDkTanThuGiaHanModel {
  iddangkytanthu: string;
  truluongconlai: number;
  thoihankhaithac: number;
  giahandenngay: string;
  donvitruluong: string;
  donvithoihan: string;
  lydogiahan: string;
  idhoso: string;
  idgiayphep: string;
  sogiayphep: string;
  ngaycapphep: string;
  siteid: string;
}

export class InputDkTanThuGiaHanModel {
  iddangkytanthu: string;
  truluongconlai: number;
  thoihankhaithac: number;
  giahandenngay: string;
  donvitruluong: string;
  donvithoihan: string;
  lydogiahan: string;
  idhoso: string;
  idgiayphep: string;
  siteid: string;
}